Work towards fixing #277 - AutoUpdate: Interrupted startup process occurs occasionally.
The auto-update process has been moved to run inside an IHostedService so that it isn't interrupted by anything the request may be doing. The existing AutoUpdateMiddleware has now been repurposed to react based on the status of the hosted service and wait a short time for it to complete. A new setting "Cofoundry:AutoUpdate:RequestWaitForCompletionTimeInSeconds"